Add updated submit interactive script
Added a script which submits an interactive job without creating a job file. I think this is not the optimal way to accomplish this however. An issue arose in my testing where a keyboard interrupt of the job submission would lead to an python traceback and exit while not actually removing the job from the queue.  The shell script which runs this python script has temporarily been named submit-intjob2.sh while bugs are worked out.